requirements.txt
eth-account>=0.13.4
python-dotenv>=1.1.1
websockets>=12.0